Update translator post review
